Remove default paging parameter value of 25 and support null

+++ common/src/main/java/com/gentics/mesh/core/data/page/impl/WrappedPageImpl.java
@@ -22,7 +22,7 @@ public class WrappedPageImpl<T> implements Page<T> {
 	}
 
 	@Override
-	public int getPerPage() {
+	public Long getPerPage() {
 		return wrappedPage.getPerPage();
 	}
